Added benchmarking, reworked single file project into multiple files
