+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rusty-snake"
+name = "rusty_snake"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,3 +7,10 @@ edition = "2021"
 minifb = "0.24"
 rand = "0.9.2"
 clap = { version = "4.4", features = ["derive"] }
+
+[dev-dependencies]
+criterion = "0.5"
+
+[[bench]]
+name = "food"
+harness = false
